+++ app/store/ui/darkModeStore.ts
@@ -1,16 +1,15 @@
-import { create } from "zustand";
-import { devtools, persist } from "zustand/middleware";
+import { create } from "zustand"
+import { devtools, persist } from "zustand/middleware"
 
 interface DarkModeStore {
-  isDarkMode: boolean;
-  toggleDarkMode: () => void;
+  isDarkMode: boolean
+  toggleDarkMode: () => void
 }
 
-
-type SetState = (fn: (prevState: DarkModeStore) => DarkModeStore) => void;
+type SetState = (fn: (prevState: DarkModeStore) => DarkModeStore) => void
 
 export const toggleDarkMode = (darkMode: DarkModeStore) => {
-  return darkMode.isDarkMode = !darkMode.isDarkMode
+  return (darkMode.isDarkMode = !darkMode.isDarkMode)
 }
 
 export const darkMode = (set: SetState): DarkModeStore => ({
@@ -19,12 +18,10 @@ export const darkMode = (set: SetState): DarkModeStore => ({
     set((state: DarkModeStore) => ({
       ...state,
       isDarkMode: toggleDarkMode(state),
-    }));
-  }
+    }))
+  },
 })
 
-
 const useDarkMode = create(devtools(persist(darkMode, { name: "darkMode" })))
-
 
 export default useDarkMode
