+++ next.config.js
@@ -2,9 +2,6 @@
 const nextConfig = {
   reactStrictMode: false,
   images: { domains: ["images.unsplash.com", "lh3.googleusercontent.com"] },
-  experimental: {
-    serverActions: true,
-  },
 }
 
 module.exports = nextConfig
